+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This release adds support to OneView Rest API version 600 for the hardware varia
    - SAS Logical Interconnect
    - SAS Logical Interconnect Group
    - Server Hardware
+   - Server Profile
    - Server Profile Template
 
 ## v5.3.0